+++ source/mplayer/osdep/timer.h
@@ -27,7 +27,7 @@ void InitTimer(void);
 #include <ogc/lwp_watchdog.h>
 inline static u64 GetTimer(void) {return ticks_to_microsecs(gettime());}
 inline static u64 GetTimerMS(void){return ticks_to_millisecs(gettime());}
-inline static int usec_sleep(unsigned int usec_delay){return usleep(usec_delay);}
+inline static int usec_sleep(unsigned long usec_delay){return usleep(usec_delay);}
 u64 GetRelativeTime(void);
 #else
 unsigned int GetTimer(void);
